Add a helper class for parsing a device descriptor
This allows parsing data from a buffer that the caller doesn't own.

This does unfortunately duplicate some code with the existing
DeviceDescriptor class.  We could make the DeviceDescriptor class
inherit from DeviceDescriptorParser in order to reduce some of the code.
However, this would slightly increase the size of the DeviceDescriptor
object.  All-in-all I'm more in favor of duplicating the accessor
methods rather than using inheritance here.