Minor directory restructuring
Move most device-specific code to the dev/ subdirectory.
I kept UsbDevice.h at the top-level just to help ensure it is more
obvious, even though it is device-specific.

Move the esp32 code into a hw/ subdirectory, so we have a location for
other hardware-specific platforms in the future.  (I have some other
projects using atmega32u4 chips, so I will hopefully add support for
that in the not too distant future.)