[esp32] rewrite interrupt to main task notification
Eliminate the code using FreeRTOS queues, and instead use a fixed set of
flags indicating which interrupts have occurred.  This uses less memory,
and avoids having to choose a queue size and dealing with possible queue
overflow.  This also makes it much simpler for users to provide their
own task main loop, as we now just need a simple notification to wake up
the task loop, and don't need to pass event data over this API.