[esp32] always reset device state in USB task after a reset
On a bus reset interrupt we immediately reset the USB registers in the
interrupt handler.  This diff makes us now reset the registers again in
the main USB task, in case they had already been changed again by
previous events that occurred before the reset but that hadn't finished
being processed when we received the reset.